fixed bugs on profile page

+++ lib/Screens/Profile/component/body.dart
@@ -15,9 +15,8 @@ import 'package:image_picker/image_picker.dart';
 import 'package:provider/provider.dart';
 
 class Body extends StatefulWidget {
-  const Body({
-    Key key,
-  }) : super(key: key);
+  final UserDataModel userData;
+  const Body({Key key, this.userData}) : super(key: key);
 
   @override
   _BodyState createState() => _BodyState();
@@ -32,157 +31,160 @@ class _BodyState extends State<Body> {
   File imageFile;
 
   @override
+  void initState() {
+    super.initState();
+    _nameController.text = widget.userData.displayName;
+    _phoneNumberController.text = widget.userData.phoneNumber;
+    _addressController.text = widget.userData.address;
+    _cityController.text = widget.userData.city;
+    _descriptionController.text = widget.userData.description;
+  }
+
+  @override
+  void dispose() {
+    _nameController.dispose();
+    _phoneNumberController.dispose();
+    _addressController.dispose();
+    _cityController.dispose();
+    _descriptionController.dispose();
+    super.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
-    final database = Provider.of<FirestoreService>(context);
     final user = Provider.of<FirebaseAuthService>(context).getCurrentUser();
+    String imageURL = widget.userData.photoURL;
     return MainBackground(
       child: SingleChildScrollView(
-        child: StreamBuilder<UserDataModel>(
-            stream: database.userDataStream(),
-            builder: (context, snapshot) {
-              if (snapshot.hasData) {
-                final userData = snapshot.data;
-                final String imageURL = userData.photoURL;
-                return Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Padding(
+              padding: const EdgeInsets.only(top: 10),
+              child: SizedBox(
+                height: 115,
+                width: 115,
+                child: Stack(
+                  fit: StackFit.expand,
+                  clipBehavior: Clip.none,
                   children: [
-                    Padding(
-                      padding: const EdgeInsets.only(top: 10),
+                    CircleAvatar(
+                      backgroundImage: imageFile == null
+                          ? (imageURL == "" ? null : NetworkImage(imageURL))
+                          : FileImage(imageFile),
+                    ),
+                    Positioned(
+                      left: -16,
+                      bottom: 0,
                       child: SizedBox(
-                        height: 115,
-                        width: 115,
-                        child: Stack(
-                          fit: StackFit.expand,
-                          clipBehavior: Clip.none,
-                          children: [
-                            CircleAvatar(
-                              backgroundImage: imageFile == null
-                                  ? NetworkImage(imageURL)
-                                  : FileImage(imageFile),
-                            ),
-                            Positioned(
-                              left: -16,
-                              bottom: 0,
-                              child: SizedBox(
-                                height: 46,
-                                width: 46,
-                                child: TextButton(
-                                  style: ButtonStyle(
-                                    shape: MaterialStateProperty.all<
-                                        RoundedRectangleBorder>(
-                                      RoundedRectangleBorder(
-                                        borderRadius: BorderRadius.circular(50),
-                                        side: BorderSide(color: Colors.white),
-                                      ),
-                                    ),
-                                    backgroundColor:
-                                        MaterialStateProperty.all<Color>(
-                                            Color(0xFFF5F6F9)),
-                                  ),
-                                  onPressed: () async {
-                                    imageFile = null;
-                                    setState(() {});
-                                  },
-                                  child: Icon(
-                                    Icons.refresh,
-                                    color: kPrimaryColor,
-                                  ),
-                                ),
+                        height: 46,
+                        width: 46,
+                        child: TextButton(
+                          style: ButtonStyle(
+                            shape: MaterialStateProperty.all<
+                                RoundedRectangleBorder>(
+                              RoundedRectangleBorder(
+                                borderRadius: BorderRadius.circular(50),
+                                side: BorderSide(color: Colors.white),
                               ),
                             ),
-                            Positioned(
-                              right: -16,
-                              bottom: 0,
-                              child: SizedBox(
-                                height: 46,
-                                width: 46,
-                                child: TextButton(
-                                  style: ButtonStyle(
-                                    shape: MaterialStateProperty.all<
-                                        RoundedRectangleBorder>(
-                                      RoundedRectangleBorder(
-                                        borderRadius: BorderRadius.circular(50),
-                                        side: BorderSide(color: Colors.white),
-                                      ),
-                                    ),
-                                    backgroundColor:
-                                        MaterialStateProperty.all<Color>(
-                                            Color(0xFFF5F6F9)),
-                                  ),
-                                  onPressed: () async {
-                                    final imagePicker =
-                                        Provider.of<ImagePickerService>(context,
-                                            listen: false);
-                                    imageFile = await imagePicker.pickImage(
-                                        source: ImageSource.gallery);
-                                    setState(() {});
-                                  },
-                                  child: Icon(
-                                    Icons.camera_alt,
-                                    color: kPrimaryColor,
-                                  ),
-                                ),
-                              ),
-                            )
-                          ],
+                            backgroundColor: MaterialStateProperty.all<Color>(
+                                Color(0xFFF5F6F9)),
+                          ),
+                          onPressed: () async {
+                            imageFile = null;
+                            setState(() {});
+                          },
+                          child: Icon(
+                            Icons.refresh,
+                            color: kPrimaryColor,
+                          ),
                         ),
                       ),
                     ),
-                    buildName(user, userData),
-                    SizedBox(height: 25),
-                    InputFieldWithTitle(
-                      title: "Display Name",
-                      content: userData.displayName,
-                      icon: Icons.person,
-                      controller: _nameController,
-                    ),
-                    InputFieldWithTitle(
-                      title: "Phone Number",
-                      content: userData.phoneNumber,
-                      icon: Icons.phone_android,
-                      controller: _phoneNumberController,
-                    ),
-                    InputFieldWithTitle(
-                      title: "Address",
-                      content: userData.address,
-                      icon: Icons.home,
-                      controller: _addressController,
-                    ),
-                    InputFieldWithTitle(
-                      title: "City",
-                      content: userData.city,
-                      icon: Icons.location_city,
-                      controller: _cityController,
-                    ),
-                    InputFieldWithTitle(
-                      title: "Description",
-                      content: userData.description,
-                      maxlines: 4,
-                      icon: Icons.description,
-                      controller: _descriptionController,
-                    ),
-                    SizedBox(height: 25),
-                    RoundedButton(
-                      text: "Save Changes",
-                      press: () {
-                        editUserData(
-                            context,
-                            _nameController.text.trim(),
-                            _phoneNumberController.text.trim(),
-                            _addressController.text.trim(),
-                            _cityController.text.trim(),
-                            _descriptionController.text.trim(),
-                            imageFile);
-                      },
-                    ),
-                    SizedBox(height: 25),
+                    Positioned(
+                      right: -16,
+                      bottom: 0,
+                      child: SizedBox(
+                        height: 46,
+                        width: 46,
+                        child: TextButton(
+                          style: ButtonStyle(
+                            shape: MaterialStateProperty.all<
+                                RoundedRectangleBorder>(
+                              RoundedRectangleBorder(
+                                borderRadius: BorderRadius.circular(50),
+                                side: BorderSide(color: Colors.white),
+                              ),
+                            ),
+                            backgroundColor: MaterialStateProperty.all<Color>(
+                                Color(0xFFF5F6F9)),
+                          ),
+                          onPressed: () async {
+                            final imagePicker = Provider.of<ImagePickerService>(
+                                context,
+                                listen: false);
+                            imageFile = await imagePicker.pickImage(
+                                source: ImageSource.gallery);
+                            setState(() {});
+                          },
+                          child: Icon(
+                            Icons.camera_alt,
+                            color: kPrimaryColor,
+                          ),
+                        ),
+                      ),
+                    )
                   ],
-                );
-              } else if (snapshot.hasError) {
-                return Text("No data available");
-              }
-              return Center(child: CircularProgressIndicator());
-            }),
+                ),
+              ),
+            ),
+            buildName(user, widget.userData),
+            SizedBox(height: 25),
+            InputFieldWithTitle(
+              title: "Display Name",
+              icon: Icons.person,
+              controller: _nameController,
+            ),
+            InputFieldWithTitle(
+              title: "Phone Number",
+              icon: Icons.phone_android,
+              controller: _phoneNumberController,
+            ),
+            InputFieldWithTitle(
+              title: "Address",
+              icon: Icons.home,
+              controller: _addressController,
+            ),
+            InputFieldWithTitle(
+              title: "City",
+              icon: Icons.location_city,
+              controller: _cityController,
+            ),
+            InputFieldWithTitle(
+              title: "Description",
+              maxlines: 4,
+              icon: Icons.description,
+              controller: _descriptionController,
+            ),
+            SizedBox(height: 25),
+            RoundedButton(
+              text: "Save Changes",
+              press: () {
+                editUserData(
+                    context,
+                    _nameController.text.trim(),
+                    _phoneNumberController.text.trim(),
+                    _addressController.text.trim(),
+                    _cityController.text.trim(),
+                    _descriptionController.text.trim(),
+                    imageURL,
+                    imageFile);
+              },
+            ),
+            SizedBox(height: 25),
+          ],
+        ),
       ),
     );
   }
@@ -192,14 +194,12 @@ class InputFieldWithTitle extends StatelessWidget {
   const InputFieldWithTitle(
       {Key key,
       @required this.title,
-      @required this.content,
       @required this.controller,
       this.maxlines,
       this.icon})
       : super(key: key);
 
   final String title;
-  final String content;
   final TextEditingController controller;
   final int maxlines;
   final IconData icon;
@@ -217,7 +217,6 @@ class InputFieldWithTitle extends StatelessWidget {
           maxLines: maxlines,
           icon: icon,
           controller: controller,
-          hintText: content,
         ),
       ],
     );
@@ -255,6 +254,7 @@ Future<void> editUserData(
     String address,
     String city,
     String description,
+    String imageUrl,
     File imageFile) async {
   try {
     String downloadUrl = "";
@@ -273,10 +273,13 @@ Future<void> editUserData(
         address: address,
         city: city,
         description: description,
-        photoURL: downloadUrl);
+        photoURL: imageUrl);
+    if (downloadUrl != "") {
+      userData.photoURL = downloadUrl;
+    }
     final database = Provider.of<FirestoreService>(context, listen: false);
     await database.setUserData(userData);
-    await imageFile.delete();
+    if (imageFile != null) await imageFile.delete();
   } catch (e) {
     print(e);
   }
